specify dtypes

+++ pandas/tests/apply/test_numba.py
@@ -44,7 +44,9 @@ def test_numba_vs_python_reductions(float_frame, reduction, apply_axis):
 @pytest.mark.parametrize("colnames", [[1, 2, 3], [1.0, 2.0, 3.0]])
 def test_numba_numeric_colnames(colnames):
     # Check that numeric column names lower properly and can be indxed on
-    df = DataFrame(np.array([[1, 2, 3], [4, 5, 6], [7, 8, 9]]), columns=colnames)
+    df = DataFrame(
+        np.array([[1, 2, 3], [4, 5, 6], [7, 8, 9]], dtype=np.int64), columns=colnames
+    )
     first_col = colnames[0]
     f = lambda x: x[first_col]  # Get the first column
     result = df.apply(f, engine="numba", axis=1)
